feat: autor crud endpoints

+++ src/main/java/com/lhcamposs/Biblioteca_API/repository/AutorRepository.java
@@ -3,5 +3,5 @@ package com.lhcamposs.Biblioteca_API.repository;
 import com.lhcamposs.Biblioteca_API.entity.Autor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AutorRepository extends JpaRepository<Autor, Integer> {
+public interface AutorRepository extends JpaRepository<Autor, Long> {
   }
