+++ src/main/java/com/lhcamposs/Biblioteca_API/repository/EditoraRepository.java
@@ -3,5 +3,5 @@ package com.lhcamposs.Biblioteca_API.repository;
 import com.lhcamposs.Biblioteca_API.entity.Editora;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EditoraRepository extends JpaRepository<Editora, Integer> {
+public interface EditoraRepository extends JpaRepository<Editora, Long> {
 }
